backend/user_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic[email]==2.6.3
motor==3.3.2
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
pymongo==4.6.1
bcrypt==4.1.2
email-validator==2.1.0.post1
python-jose==3.3.0